Support multiple filters, fix bugs and clean up.
Adds support for multiple filter (-f) options so that several criteria
to save histograms can be applied efficiently in a single pass over
the data.

Fixes handling of filter patterns ending in a slash. Detects such
patterns correctly as matching a directory entity, and avoids
unnecessary descent into the directory once it has been matched.

Avoids labelling intermediate directories as matches. Any needed
intermediate directories dynamically as needed, if and only if writing
an output root file. While scanning directories are reported as search
matches if and only if the pattern matches a directory: the pattern
ends in a trailing slash and data contains a monitor element in such a
directory.

Cleans up white space and otherwise tidies up the code a bit.

Signed-off-by: Marco Rovere <marco.rovere@cern.ch>


git-svn-id: svn+ssh://svn.cern.ch/reps/CMSDMWM/Monitoring/trunk@15760 4525493e-7705-40b1-a816-d608a930855b