Upgraded Index schema: the number of available bits used to assemble the keys internally used has been increased from 64 to 128 bits. The version number of the index has been bumped by 1 unit. Indexes created with the new schema are not backward compatible with the previous version. In particular all binary executables will be bound to a specific schema and will refuse to operate on a different version(i.e. on a different schema). This implies that care must be taken in order to properly operate a server: the correct environemnt should be sourced in order to couple the index schema and the version of the binaries.
VisDQMImportDaemon is now able to index also the intermediate .dat files. The
ordering is such that .dat files come first with respect to ordinary root
files: this choice is motivated by the fact that we want to give highest
priority to index-migration jobs. Within .dat files, the ordering is the same
as the one used to order .root files.