CastorVerifier: Fixed recovery corrupt pickle file
We had a problem with this when the hard drive was completely filled
during the incident with the dev server on 30/06/2015.
I noticed now that in fact there was already code to catch the situation
in which the pickle file is corrupted. But there was a problem with the
way it tried to recover.