Renaming Quota/VerControl, removing AfsSync
For some reason visDQMRootFileQuotaControl didn't end with daemon
anymore, while all the other deamons do. (I probably did this myself)
This has implications in the stop methods in the manage script and it's
just simply not consistent.
If we have "RootFileQuotaControl", then "VerControl" should also be
called "RootFileVersionControl".
And finaly, removed the visDQMAfsSync (which was also a daemon btw).
Even though this code might come in handy some day, it's not a reason to
keep it. If we need it we can get it from the history.