Add Upload performance test
A recent upload-storm to the production servers, revealed that the way
in which cherry-py natively support file-upload via
multipart/form-data is suboptimal. A patch has been provided locally
to our cherry-py distribution in order to speed up the transfer in all
cases in which each part of the multipart has explicitely specified
the Content-length header. This test is to verify that the performance
of the file-upload with and without the content-length specification
indeed differ a lot. This should be monitored at every cherry-py
upgrade, to keep the upload performances under control.