Fix checksum on "binary" imported modules.
The way in which the inspect module works changed and, in case we
import a "binary" module, like our Accelerator one, an exception is
thrown even if the source code of that module is present. This PR is a
quick and dirty, relatively clean way, to go around the problem. With
this, the server is happily starting. The proper fix, most likely,
should be implemented in the buggy inspect module.

Also, the import_module local function has been dropped and the one
imported from the importlib module is used.

fix