Reformat the header appearence
The font size of the header are not expressed using em units to allow for a
better/automatic scaling on different browser settings and screen sizes. Since
the font size has also been increase, few adjustments have been made to the
spacing of the different components of the header menu-bar. Pure stylistical,
no functionality added whatsoever.