Add TF1 to the list of StreamerInfo
Some TH1 objects apparently are shipped to the DQMGUI together with some
associated function. The schema evolution of such objects will not be
guaranteed unless their streamerInfo is saved internally to our
database. This PR will automatically trigger the creation of a new
StreamerInfo entry in the master table that will contain also the TF1
information. All patches that need to be applied to an existing index/DB
**must be done before inserting anything into the DB**, otherwise the
effect of the patch will be null (the patch will only affect the last
entry in the StreamerInfo table, since that's the only one guaranteed to
be in sync with the correct version of ROOT). Also, the patch must be
applied keeping the same version of ROOT that was used to produce the
last entry in the StreamerInfo table, otherwise the DB will be
corrupted.