Patch Streamers different from last
In the past only the last entry of the streamerInfo table
could be patched. Since a use case raised to be able to
patch the last-but-one entry, the code has been modified in
such a way that any arbitrary entry can be patched. The
default entry, if none is supplied, is still the last one.
We fully rely on the end user to know exactly what she is
doing. Some other minor improvement in debugging and
debugging messages has been stuffed in.