Modify Makefile to build against newer code from CMSSW
Newer code from CMSSW was needed in order to be able to
compile DQMGUI code against ROOT 6.06. This required the
addition of few dependencies for DQMStore that had to be
taken care of in the Makefile. Also, always remember to
compile(protoc) all .proto messages that could come into
from CMSSW, since we have no guarantee that the generated
code from CMSSW will be compatible with DQMGUI
ProtocolBuffer library.