Protect against weird image parameters
Absolute image sizes are regularly checked before trying to
produce an image. An additional check has been added in the
case in which the original width and height were within the
allowed limit but in a ratio such that a rescaling was
forced, whose output was not checked. This caused few
crashes in the rendering processes that caused the main
webserver to restart.