Filter input on the search box
Users are free to search for samples registered in the
DQMGUI backend using a free-text regular expression field. A
check is now made on the client side to ensure that the
input only contains valid characters for a regular
expression, avoiding unicode queries that are not supported
on the server side.