Set custom range for references
In ROOT, when a histogram is drawn with a user-defined
axis-range, the content of the stat-box is updated to
consider only the entries within the selected range. When an
additional histogram is overlaid on top, the user-range
selection is not applied, so that the stat box shows the
statistics coming from the full sample, making the
comparison with the original histogram difficult, if not
impossible. DQMGUI will therefore impose the user-range also
to reference histograms, to force to have them the
statistics updated properly.