Dynamic Y-range for overlaid histograms
Add the automatic computation of the Y-range in case of overlaid
histograms. The final logic is such that the manual parameters that the
user could have set have precedence over the automatic value, i.e. the
manual user range is applied last. One loop over all references is
needed in order to pre-compute the optimal range before drawing
anything. The optimal range takes into account also the normalization of
the reference plots. Reference plots that the user decides not to
normalize are not considered by the automatic algorithm. Small
refactoring of the code to avoid unecessary casts.