Limit automatic riscaling to overlaid plots
The automatic rescaling has been limited to overlaid plots (i.e.
the ones that have an active rerefence). This to avoid weird
behaviour with colors and render-plugins for 2D histograms whose
range is fixed in the plugin to conform to some pre-formatted
color palette, that the automatic rescaling was basically
scrambling. This will hide the problem: a rescale manully done
within the DQMGUI on these plots will in any case result in
weird/wrong color coding. The permanent solution would be to
assign a particular color to a specific range, regardless of the
Y{Z}-range extension, but at present this feature is not
available in ROOT.