Limit automatic rescaling to 1D plots
The automatic rescaling is playing really bad with the logic
implemented in most render-plugins to assign specific colors
to specific intervals.  Since the automatic rescaling will
change the Y/Z range, the assignment will be broken. To
avoid this problem completely, the rescaling is only
triggerd for 1D plots, where we do not have this color
assignment issue.