Speed up rendering
In cases in which a plots is drawn with one or more
references, few pstacks of the rendering processes showed
that most of the time was spent in drawing each and every
marker of the reference plots. We therefore dropped the
marker support for  reference histograms completely. The
gain in timing was quite impressive.