Protect JSON interface from weird numbers
Recently we experienced a huge number of crashes related to
the JSON API. The crashes were caused by an interplay
between ridicolously small/big numbers and a wrong
formatting options for them. We used a more natural 'g'
formatting for all cases, so that even extremely tiny or big
numbers are properly formatted. The origin of the crashes is
documented at this URL:
https://hypernews.cern.ch/HyperNews/CMS/get/dqmDevel/2357.html
Also, a proper handling of inf values has been added to the
final JSON formatting, in order to always receive a valid
JSON that can be parsed and represented in the DQMGUI.