Add TAxis to the list of saved streamerinfos
The TAxis component of a histogram was not natively saved
into DQMGUI's index and that caused problems the very first
moment the TAxis went through a schema evolution. This PR
adds the TAxis to the list of saved streamerinfos. The way
to deliver this change is to first deploy a version of
DQMGUI with this change in, then patch the index to modify
the old streamerinfo to include the newly added TAxis, then
start the DQMGUI web-service. Failing to follow this
procedure will likely make old samples unreadable.