Protect streamerInfo patch
Trying to patch the streamerInfo stored in the current
database using a novel streamerInfo that is identical to any
of the ones already present would result in a corrupted
index. A protection is added to check that the patched
streamerInfo has never ever been used. If that's not the
case, the patch will terminate with an assert, leaving the
index untouched and usable.