Rewrite build streamer info logic
In the past, DQMGUI suffered from schema evolution of
objects that were not directly TH* or TProfile, but that
were somehow needed to properly decode the stored streamed
information. As a temporary fix, the missing streamerInfo
was manually added to the index, patching it. But this
guaranteed a compatibility only with the samples produced
with the N-1 streamer, N being the new one. Old samples
suffer the schema evolution problem. In order to permanently
fix this issue, a full list of required classes is derived
starting from the ones really used inside the uploaded ROOT
files, taking into account class inheritance and data
members' type.