Improve protection against weird images
Removed the assert statement on the input file size since it
was too strict and was causing repeated crashes of the web
application. The assert has been dropped in favour of a
completely black image in case the assert condition is not
satisfied: this will alert the user/developer that there is
something weird with the plot. The problem of the crashes
was due to a buggy DQMGenericClient that filled some
efficiency plots with nans that completely screw the drawing
process.